diary /diary v2/backend/app/main.py
import os
import logging

from fastapi import FastAPI, Request, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
from fastapi.responses import JSONResponse
from fastapi.exceptions import RequestValidationError
from starlette.exceptions import HTTPException as StarletteHTTPException

from .api import auth, entries, insights, transcribe
from .core.config import get_settings

logger = logging.getLogger(__name__)

settings = get_settings()
app = FastAPI(title="Voice Journal API", version="0.1.0")

# CORS middleware must be added before exception handlers
app.add_middleware(
    CORSMiddleware,
    allow_origins=settings.allowed_origins or ["http://localhost:5173"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

if not settings.media_base_url.startswith("http"):
    settings.media_root.mkdir(parents=True, exist_ok=True)
    app.mount(settings.media_base_url, StaticFiles(directory=settings.media_root), name="media")

# Global exception handlers to ensure CORS headers are always included
@app.exception_handler(StarletteHTTPException)
async def http_exception_handler(request: Request, exc: StarletteHTTPException):
    """Handle HTTP exceptions with CORS headers."""
    return JSONResponse(
        status_code=exc.status_code,
        content={"detail": exc.detail},
        headers={
            "Access-Control-Allow-Origin": request.headers.get("origin", "http://localhost:5173"),
            "Access-Control-Allow-Credentials": "true",
        },
    )


@app.exception_handler(RequestValidationError)
async def validation_exception_handler(request: Request, exc: RequestValidationError):
    """Handle validation errors with CORS headers."""
    return JSONResponse(
        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
        content={"detail": exc.errors()},
        headers={
            "Access-Control-Allow-Origin": request.headers.get("origin", "http://localhost:5173"),
            "Access-Control-Allow-Credentials": "true",
        },
    )


@app.exception_handler(Exception)
async def general_exception_handler(request: Request, exc: Exception):
    """Handle all other exceptions with CORS headers."""
    logger.exception("Unhandled exception", exc_info=exc)
    # Include error message in development for debugging
    error_detail = str(exc) if os.getenv("ENVIRONMENT", "development") != "production" else "Internal server error"
    return JSONResponse(
        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
        content={"detail": error_detail},
        headers={
            "Access-Control-Allow-Origin": request.headers.get("origin", "http://localhost:5173"),
            "Access-Control-Allow-Credentials": "true",
        },
    )


app.include_router(auth.router)
app.include_router(entries.router)
app.include_router(entries.tag_router)
app.include_router(insights.router)
app.include_router(transcribe.router)


@app.get("/healthz")
def health_check():
    return {"status": "ok"}
