diary /diary v2/backend/pyproject.toml
[project]
name = "voice-journal-backend"
version = "0.1.0"
description = "FastAPI backend for the voice journal MVP"
authors = [{name = "Voice Journal"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<5.0.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "aiofiles>=24.1.0",
  "httpx>=0.27.0",
  "openai>=1.45.0",
  "email-validator>=2.3.0",
  "ffmpeg-python>=0.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.5.7"
]

[tool.ruff]
line-length = 100
exclude = ["alembic"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
